refactor(cognition): split model_resolver.rs (1232 LOC) into mod + types (#1208)
`cognition/model_resolver.rs` was 1232 LOC mixing eight type
definitions with the resolver function and its tests. Split:

  model_resolver/mod.rs    (939 LOC) — derive_target_silicon helper +
                                       resolve_model function +
                                       full test suite (20 tests
                                       including 5-persona smoke +
                                       standard-persona alpha bar tests)
  model_resolver/types.rs  (321 LOC) — HwCapabilityTier (16 variants),
                                       SiliconResidencyRequirement +
                                       allows() impl, LocalOrCloudPolicy,
                                       HostCapability, ModelRequirement +
                                       standard_persona constructors,
                                       ResolvedModel, ResolutionError
                                       (3 variants with thiserror messages),
                                       7 ts-rs export tests

All types re-exported from `mod.rs` via `pub use types::{...}` — external
callers of resolver types see no API change.

Tests:
- `cargo check --features metal,accelerate -p continuum-core` clean.
- `cargo test --features metal,accelerate -p continuum-core --lib model_resolver`
  → 27 passed, 0 failed (20 resolver + 7 ts-rs bindings).

Closes #1208 for the model_resolver.rs slice. evaluator.rs split
shipped in #1242. admission.rs (1225) remains deferred — its tests
are tightly woven through AdmissionGate+IsMemorable composition;
separate card.
